seperate renderer with chart controlling
